platform/analyzer/requirements.txt
# Flask-related
click==8.1.3
Flask~=2.2.3
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.2
pip==23.0.1
setuptools==67.6.0
Werkzeug==2.2.3
wheel==0.40.0

# Misc.
openai~=0.27.2
python-dotenv~=1.0.0
gunicorn~=20.1.0
requests~=2.28.2

# Geo
geograpy3~=0.2.5
geopy~=2.3.0
pycountry~=22.3.0

# NLP
# Also this is required: python -m spacy download en_core_web_sm
spacy~=3.5
